Switch float64 to float32 and experimental 3D radianca cascade

+++ src/2D/main.go
@@ -10,7 +10,7 @@ import (
 	"strings"
 )
 
-func NewScene(sceneAsString string, time float64) scene.Scene {
+func NewScene(sceneAsString string, time float32) scene.Scene {
 	sceneAsString = strings.ToLower(sceneAsString)
 	var s scene.Scene = nil
 	switch sceneAsString {
