+++ src/2D/render/radiance_cascade/bilinear_fix.go
@@ -6,18 +6,18 @@ import (
 
 func GetBilinearFixProbe(ciNear *CascadeInfo, ciFar *CascadeInfo, iNear int, jNear int, iFar int, jFar int, index int) CascadeProbe {
 
-	angle := ciNear.angleStart + ciNear.deltaAngle*float64(index)
+	angle := ciNear.angleStart + ciNear.deltaAngle*float32(index)
 	dir := vector.NewVec2fromAngle(angle)
 	tstart := ciNear.TStart
 	tend := ciNear.TEnd
 
 	pNear := vector.Vec2{
-		X: ciNear.p0.X + float64(iNear)*ciNear.spacing + dir.X*tstart,
-		Y: ciNear.p0.Y + float64(jNear)*ciNear.spacing + dir.Y*tstart,
+		X: ciNear.p0.X + float32(iNear)*ciNear.spacing + dir.X*tstart,
+		Y: ciNear.p0.Y + float32(jNear)*ciNear.spacing + dir.Y*tstart,
 	}
 	pFar := vector.Vec2{
-		X: ciFar.p0.X + float64(iFar)*ciFar.spacing + dir.X*tend,
-		Y: ciFar.p0.Y + float64(jFar)*ciFar.spacing + dir.Y*tend,
+		X: ciFar.p0.X + float32(iFar)*ciFar.spacing + dir.X*tend,
+		Y: ciFar.p0.Y + float32(jFar)*ciFar.spacing + dir.Y*tend,
 	}
 	pFar.Sub(pNear)
 	length := pFar.Normalize()
@@ -37,9 +37,9 @@ func (rc *RadianceCascade) CascadeMergeBilinearFix(cNear *Cascade, cFar *Cascade
 	ciNear := cNear.info
 	ciFar := cFar.info
 
-	factor := float64(ciNear.DirCount) / float64(ciFar.DirCount) // integration factor
+	factor := float32(ciNear.DirCount) / float32(ciFar.DirCount) // integration factor
 	nDirMerge := ciFar.DirCount / ciNear.DirCount                // number of directions to merge
-	bilinearWeights := vector.Vec2{X: 0.25 + 0.5*float64(x&1), Y: 0.25 + 0.5*float64(y&1)}
+	bilinearWeights := vector.Vec2{X: 0.25 + 0.5*float32(x&1), Y: 0.25 + 0.5*float32(y&1)}
 	merged := NewCascadeRadianceResult()
 	if ciNear.isLast {
 		cNear.radiance[x][y][k] = rc.Radiance(ciNear.GetProbe(x, y, k))
